create setCookies as a  separate file
